Extract server handling logic to server package
